src/requirements.txt
pandas
requests
BeautifulSoup4
selenium
lxml
python-dotenv
pymssql